update to latest SQLKit 2 changes
