+++ Sources/PostgresKit/Exports.swift
@@ -1,0 +1,2 @@
+@_exported import NIOPostgres
+@_exported import DatabaseKit
