+++ Sources/PostgresKit/PostgresDatabase.swift
@@ -77,15 +77,15 @@ public final class PostgresDatabase: Database {
         do {
             address = try self.config.address()
         } catch {
-            return self.eventLoop.makeFailedFuture(error: error)
+            return self.eventLoop.makeFailedFuture(error)
         }
-        return PostgresConnection.connect(to: address, on: self.eventLoop).then { conn in
+        return PostgresConnection.connect(to: address, on: self.eventLoop).flatMap { conn in
             return conn.authenticate(
                 username: self.config.username,
                 database: self.config.database,
                 password: self.config.password
             ).map { conn }
-        }.then { conn in
+        }.flatMap { conn in
             if let tlsConfig = self.config.tlsConfig {
                 return conn.requestTLS(using: tlsConfig).map { upgraded in
                     if !upgraded {
@@ -95,7 +95,7 @@ public final class PostgresDatabase: Database {
                     return conn
                 }
             } else {
-                return self.eventLoop.makeSucceededFuture(result: conn)
+                return self.eventLoop.makeSucceededFuture(conn)
             }
         }
     }
@@ -114,36 +114,72 @@ extension PostgresRow: SQLRow {
     public func decode<D>(_ type: D.Type, table: String?) throws -> D
         where D: Decodable
     {
-        if let table = table {
-            return try self.decode(D.self, table: table)
-        } else {
-            return try self.decode(D.self, tableOID: 0)
+        #warning("TODO: implement decoding")
+        fatalError()
+    }
+}
+
+public struct SQLRaw: SQLExpression {
+    public var string: String
+    public init(_ string: String) {
+        self.string = string
+    }
+    
+    public func serialize(to serializer: inout SQLSerializer) {
+        serializer.write(self.string)
+    }
+}
+
+struct PostgresDialect: SQLDialect {
+    private var bindOffset: Int
+    
+    init() {
+        self.bindOffset = 0
+    }
+    
+    var identifierQuote: SQLExpression {
+        return SQLRaw("\"")
+    }
+    
+    var literalStringQuote: SQLExpression {
+        return SQLRaw("'")
+    }
+    
+    mutating func nextBindPlaceholder() -> SQLExpression {
+        self.bindOffset += 1
+        return SQLRaw("$" + self.bindOffset.description)
+    }
+    
+    func literalBoolean(_ value: Bool) -> SQLExpression {
+        switch value {
+        case false:
+            return SQLRaw("false")
+        case true:
+            return SQLRaw("true")
         }
     }
     
-    
+    var autoIncrementClause: SQLExpression {
+        return SQLRaw("GENERATED BY DEFAULT AS IDENTITY")
+    }
 }
 
 extension PostgresConnection: SQLDatabase {
-    public func execute(_ query: PostgresQuery, _ onRow: @escaping (PostgresRow) throws -> ()) -> EventLoopFuture<Void> {
-        var binds: [Encodable] = []
-        let sql = query.serialize(&binds)
-        var b = PostgresBinds()
-        binds.forEach { b.encode($0) }
-        return self.query(sql, b, onRow).then {
-            switch query.storage {
-            case .alterTable, .createTable, .dropTable:
-                return self.loadTableNames()
-            default:
-                return self.eventLoop.makeSucceededFuture(result: ())
-            }
+    public func execute(_ query: SQLExpression, _ onRow: @escaping (SQLRow) throws -> ()) -> EventLoopFuture<Void> {
+        var serializer = SQLSerializer(dialect: PostgresDialect())
+        query.serialize(to: &serializer)
+        print(serializer.sql)
+        return self.query(serializer.sql, serializer.binds.map { encodable in
+            return PostgresData.null
+        }) { row in
+            try onRow(row)
         }
     }
 }
 
 extension PostgresDatabase: SQLDatabase {
-    public func execute(_ query: PostgresQuery, _ onRow: @escaping (PostgresRow) throws -> ()) -> EventLoopFuture<Void> {
-        return self.newConnection().then { conn in
+    public func execute(_ query: SQLExpression, _ onRow: @escaping (SQLRow) throws -> ()) -> EventLoopFuture<Void> {
+        return self.newConnection().flatMap { conn in
             return conn.execute(query, onRow)
         }
     }
