+++ Sources/PostgresKit/Query/PostgresQuery+Collation.swift
@@ -1,11 +1,11 @@
-import SQLKit
-
-extension PostgresQuery {
-    #warning("implement me")
-    public struct Collation: SQLCollation {
-        public func serialize(_ binds: inout [Encodable]) -> String {
-            return "COLLATE"
-        }
-    }
-}
-
+//import SQLKit
+//
+//extension PostgresQuery {
+//    #warning("implement me")
+//    public struct Collation: SQLCollation {
+//        public func serialize(_ binds: inout [Encodable]) -> String {
+//            return "COLLATE"
+//        }
+//    }
+//}
+//
