+++ Sources/PostgresKit/Query/PostgresQuery+DropIndex.swift
@@ -1,9 +1,9 @@
-import SQLKit
-
-extension PostgresQuery {
-    public struct DropIndex: SQLDropIndex {
-        public func serialize(_ binds: inout [Encodable]) -> String {
-            return .init()
-        }
-    }
-}
+//import SQLKit
+//
+//extension PostgresQuery {
+//    public struct DropIndex: SQLDropIndex {
+//        public func serialize(_ binds: inout [Encodable]) -> String {
+//            return .init()
+//        }
+//    }
+//}
