+++ Tests/PostgresKitTests/Utilities.swift
@@ -10,9 +10,8 @@ extension PostgresConnection {
             #else
             address = try .init(ipAddress: "127.0.0.1", port: 5432)
             #endif
-            return connect(to: address, on: eventLoop).then { conn in
+            return connect(to: address, on: eventLoop).flatMap { conn in
                 return conn.authenticate(username: "vapor_username", database: "vapor_database", password: "vapor_password")
-                    .then { conn.loadTableNames() }
                     .map { conn }
             }
         } catch {
